requirements.txt
discord.py
openai
tiktoken
python-dotenv
ollama